+++ lib/src/test/java/gdv/xport/DatenpaketTest.java
@@ -19,10 +19,7 @@
 package gdv.xport;
 
 import gdv.xport.config.Config;
-import gdv.xport.feld.Bezeichner;
-import gdv.xport.feld.ByteAdresse;
-import gdv.xport.feld.Datum;
-import gdv.xport.feld.Feld;
+import gdv.xport.feld.*;
 import gdv.xport.satz.*;
 import gdv.xport.satz.model.Satz100;
 import gdv.xport.util.SatzFactory;
@@ -144,6 +141,15 @@ public final class DatenpaketTest {
         assertEquals(1, nachsatz.getAnzahlSaetze());
     }
 
+    @Test
+    public void testAddVorbelegung() {
+        datenpaket.setVuNummer("007");
+        datenpaket.setVermittler("JamesBond");
+        datenpaket.add(SatzFactory.getDatensatz(SatzTyp.of(100)));
+        assertEquals(datenpaket.getVuNummer(), datenpaket.getDatensaetze().get(0).getVuNummer());
+        assertEquals(datenpaket.getVermittler(), datenpaket.getDatensaetze().get(0).getVermittler());
+    }
+
     /**
      * Falls kein Datum gesetzt wird, sollte als Default das heutige DAtum
      * zurueckgegeben werden.
@@ -171,13 +177,28 @@ public final class DatenpaketTest {
      */
     @Test
     public void testSetVermittler() {
+        Datenpaket dp = SatzRegistry.getInstance().getAllSupportedSaetze();
         String vermittler = "08/15";
-        datenpaket.setVermittler(vermittler);
-        assertEquals(vermittler, datenpaket.getVermittler());
-        Vorsatz vorsatz = datenpaket.getVorsatz();
+        dp.setVermittler(vermittler);
+        assertEquals(vermittler, dp.getVermittler());
+        Vorsatz vorsatz = dp.getVorsatz();
         assertEquals(vermittler, vorsatz.getVermittler());
-        Nachsatz nachsatz = datenpaket.getNachsatz();
+        Nachsatz nachsatz = dp.getNachsatz();
         assertEquals(vermittler, nachsatz.getVermittler());
+        for (Datensatz satz : dp.getDatensaetze()) {
+            assertEquals(vermittler, satz.getVermittler());
+        }
+    }
+
+    @Test
+    public void testSetVuNummer() {
+        Datenpaket dp = SatzRegistry.getInstance().getAllSupportedSaetze();
+        dp.setVuNummer("12345");
+        for (Datensatz satz : dp.getDatensaetze()) {
+            if (satz.hasFeld(Bezeichner.VU_NUMMER)) {
+                assertEquals("12345", satz.getVuNummer());
+            }
+        }
     }
 
     /**
@@ -185,9 +206,7 @@ public final class DatenpaketTest {
      *
      * @throws IOException Signals that an I/O exception has occurred.
      */
-    @IntegrationTest
     @Test
-    @SkipTestOn(property = "SKIP_IMPORT_TEST")
     public void testImportFromReader() throws IOException {
         try (InputStream istream = this.getClass().getResourceAsStream("/musterdatei_041222.txt")) {
             checkImport(datenpaket, istream);
@@ -281,7 +300,7 @@ public final class DatenpaketTest {
             }
         }
 
-        assertTrue("There are teildatensaetze with wrong satznummern: (datensatz, teildatensatz) -> " + wrongDatensatzTeildatensatzList.toString(), wrongDatensatzTeildatensatzList.isEmpty());
+        assertTrue("There are teildatensaetze with wrong satznummern: (datensatz, teildatensatz) -> " + wrongDatensatzTeildatensatzList, wrongDatensatzTeildatensatzList.isEmpty());
 
         assertEquals("BRBRIENNEE,J\u00dcRGEN", datenpaket.getAdressat());
     }
@@ -358,7 +377,7 @@ public final class DatenpaketTest {
         Config.setEOD("\n");
         String muster = getResourceAsString("/musterdatei_041222.txt");
         datenpaket.importFrom(muster);
-        Satz vertragsteil = datenpaket.getDatensaetze().get(2);
+        Satz vertragsteil = datenpaket.getDatensaetze().get(1);
         Feld vertragsstatus = vertragsteil.getFeld(Bezeichner.VERTRAGSSTATUS);
         assertEquals("1", vertragsstatus.getInhalt());
         checkExportWith(muster);
@@ -726,7 +745,7 @@ public final class DatenpaketTest {
     }
 
     @Test
-    public void testExportImportSchaden500() throws IOException {
+    public void testExportImportSchaden500V1_8() throws IOException {
         /*
         In der aktuellen Version (VUVM2018.xml) ist fuer Datensatz 0500 im ersten Teilsatz eine Satznummerwiederholung vorgesehen.
         Der Rueckimport muss korrekt zwei Datensaetze erkennen, selbst wenn wir jeweils den zweiten Datensatz nicht exportieren.
@@ -742,6 +761,9 @@ public final class DatenpaketTest {
         Datenpaket imported = new Datenpaket();
         imported.importFrom(testfile);
         assertEquals(2, imported.getDatensaetze().size());
+
+        Version version0500 = imported.getVorsatz().getSatzartVersionen().get(SatzTyp.of("0500"));
+        assertEquals("Version vom Datensatz 0500 stimmt nicht", "1.8", version0500 != null ? version0500.getInhalt() : "nicht vorhanden");
     }
 
     @Test
@@ -751,12 +773,11 @@ public final class DatenpaketTest {
         Wir simulieren das, indem die Satznummerwiederholung mit 'leer' geliefert wird.
         Trotzdem muss der Rueckimport korrekt zwei Datensaetze erkennen und nicht nur einen.
          */
+        SatzRegistry satzRegistry2009 = SatzRegistry.getInstance("VUVM2009.xml");
         File testfile = new File("target", "schadensatz_500_zwei_datensaetze_v1_5.gdv");
-        Datensatz schaden = SATZ_REGISTRY.getDatensatz(SatzTyp.of("0500"));
-        schaden.set("SatzNr2", " "); // setze Satznummerwiederholung auf 'leer'
+        Datensatz schaden = satzRegistry2009.getDatensatz(SatzTyp.of("0500"));
         schaden.removeTeildatensatz(2);
-        Datensatz schaden2 = SATZ_REGISTRY.getDatensatz(SatzTyp.of("0500"));
-        schaden2.set("SatzNr2", " "); // setze Satznummerwiederholung auf 'leer'
+        Datensatz schaden2 = satzRegistry2009.getDatensatz(SatzTyp.of("0500"));
         schaden2.removeTeildatensatz(2);
         datenpaket.add(schaden);
         datenpaket.add(schaden2);
@@ -764,6 +785,30 @@ public final class DatenpaketTest {
         Datenpaket imported = new Datenpaket();
         imported.importFrom(testfile);
         assertEquals(2, imported.getDatensaetze().size());
+
+        Version version0500 = imported.getVorsatz().getSatzartVersionen().get(SatzTyp.of("0500"));
+        assertEquals("Version vom Datensatz 0500 stimmt nicht", "1.5", version0500 != null ? version0500.getInhalt() : "nicht vorhanden");
+    }
+
+    /**
+     * Test dass ein Datensatz in einer alten Version korrekt ist, in einer neuen Version aber nicht korrekt
+     * Testfall eines Beispiels eines 0220.010.48-er Datensatzes, Aenderung zwischen VUVM2015 und VUVM2018 (Risikozuschlag in %)
+     *
+     * @throws IOException im Fehlerfall
+     */
+    @Test
+    public void testImport0220_010_48_version_differences() throws IOException {
+        Datenpaket datenpaket2_1 = new Datenpaket();
+        File testfile = new File("src/test/resources", "gdv/xport/satz/testcase_0220_010_48_v2_1.txt");
+        datenpaket2_1.importFrom(testfile, Charset.forName("IBM850"));
+        datenpaket2_1.validate();
+        assertTrue("Datenpaket muss gueltig sein", datenpaket2_1.isValid());
+
+        Datenpaket datenpaket2_2 = new Datenpaket();
+        testfile = new File("src/test/resources", "gdv/xport/satz/testcase_0220_010_48_v2_2_error.txt");
+        datenpaket2_2.importFrom(testfile, Charset.forName("IBM850"));
+        datenpaket2_2.validate();
+        assertFalse("Datenpaket darf nicht gueltig sein", datenpaket2_2.isValid());
     }
 
 }